Create query.txt
I started writing the queries, as I had time at work xD. When you have time check them. I also changed and the word document respectively for the task 3. As for the query e3) I wrote two versions, the first is simpler and seems ok, the second is far more complicated but more accurate to the question.